+++ docs/.vitepress/theme/index.ts
@@ -1,11 +1,13 @@
 import DefaultTheme from 'vitepress/theme'
 import type { Theme } from 'vitepress'
 import MermaidDiagram from './MermaidDiagram.vue'
+import AgentWorkbenchDemo from './AgentWorkbenchDemo.vue'
 import './custom.css'
 
 export default {
   extends: DefaultTheme,
   enhanceApp({ app }) {
     app.component('MermaidDiagram', MermaidDiagram)
+    app.component('AgentWorkbenchDemo', AgentWorkbenchDemo)
   }
 } satisfies Theme